requirements.txt
Django
psycopg2-binary
python-decouple
Pillow
djangorestframework
python-dotenv
django-cors-headers
django-filter
requests



   